Add files via upload
Code for 2048 game we play.  initially we are provided with a 4*4 matrix with all entries zero except a 2 at any random place. We can move the elements of matrix up, down, left or right according to commands mentioned. After each command all the elements present will shift in one direction and equal numbers will get combine and gets itself double. After each chance a 2 will be generated at random empty cell. At the end if we can make 2048 by combining equal numbers, then we win the game. or if all the cells of matrix gets filled and there is no empty cell left to generate a new 2 then we loss the game. Otherwise we can continue,